requirements.txt
pydub==0.23.1
piwho==1.3.2
protobuf==3.7.1
soundfile==0.10.2
google-cloud-speech==0.36.3
pyaudio==0.2.11
six==1.12.0
